Add functions to get compatible QoS profiles
Given one or more publisher (subscription) QoS profiles, modify the given
subscription (publisher) QoS profile such that it is compatible with most
of the publisher (subscription) profiles.
Do this while not sacrificing the level of service when possible.

This API is ultimately intended to be leveraged by ROS tools that
typically wants to be flexible with respect to the QoS when subscribing or
publishing to a topic.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>